+++ Bricouli/src/main/java/com/demo/Bricouli/repositories/AnnonceMRepository.java
@@ -1,11 +1,15 @@
 package com.demo.Bricouli.repositories;
 
+import java.util.List;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.demo.Bricouli.models.AnnonceMateriel;
 
 @Repository
-public interface AnnonceMRepository extends JpaRepository<AnnonceMateriel, Double> {
+public interface AnnonceMRepository extends JpaRepository<AnnonceMateriel, Long> {
+	
+	
 
 }
