+++ Bricouli/src/main/java/com/demo/Bricouli/repositories/AnnonceurRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.demo.Bricouli.models.Annonceur;
 
 @Repository
-public interface AnnonceurRepository extends JpaRepository<Annonceur, Double> {
+public interface AnnonceurRepository extends JpaRepository<Annonceur, Long> {
 
 }
